requirements.txt
python-telegram-bot==20.3
openai==1.25.0
python-dotenv==1.0.1
